Fixing AzureBatchSource to work with Filesets
