+++ core-plugins/docs/AzureBlobStore-batchsource.md
@@ -17,15 +17,13 @@ Properties
 **referenceName:** This will be used to uniquely identify this source for lineage, annotating metadata, etc.
 
 **path:** The path on Microsoft Azure Blob Storage to use as input. The path uses filename expansion (globbing) to read
-files. (Macro-enabled)
+files. The path must start with `wasb://` or `wasbs://`, for example, `wasb://mycontainer@mystorageaccount.blob.core.windows.net/filename.txt`. (Macro-enabled)
 
 **account:** The Microsoft Azure Blob Storage account to use. This is of the form 
 `mystorageaccount.blob.core.windows.net`, where `mystorageaccount` is the Microsoft 
 Azure Storage account name. (Macro-enabled)
 
-**container:** The container to use on the specified Microsoft Azure Storage account. (Macro-enabled)
-
-**storageKey:** The storage key for the specified container on the Microsoft Azure Storage account. 
+**storageKey:** The storage key for the container on the Microsoft Azure Storage account. 
 Must be a valid base64 encoded storage key provided by Microsoft Azure. (Macro-enabled)
 
 **ignoreNonExistingFolders:** Identify if path needs to be ignored or not, for case when directory or file does not
@@ -37,17 +35,15 @@ Example
 -------
 
 This example connects to Microsoft Azure Blob Storage and reads in files found in the
-specified directory. This example uses the 'hydrator' container at
-'hydratorstorage.blob.core.windows.net' (Microsoft Azure Storage), using the
-'hydratorstorage' account name:
+specified directory. This example uses Microsoft Azure Storage 'mystorageaccount.blob.core.windows.net', using the
+'mystorageaccount' account name:
 
     {
         "name": "AzureBlobStore",
         "type": "batchsource",
         "properties": {
-            "path": "/path/to/input",
-            "account": "hydratorstorage.blob.core.windows.net",
-            "container": "hydrator",
+            "path": "`wasb://mycontainer@mystorageaccount.blob.core.windows.net/filename.txt",
+            "account": "mystorageaccount.blob.core.windows.net",
             "storageKey": "XXXXXEEESSS/YYYY=",
             "ignoreNonExistingFolders": "false",
             "recursive": "false"
